rename properAddress to toBeProperAddress
